requirements.txt
watchdog>=3.0.0
transmission-rpc>=4.0.0
pathlib2>=2.3.7
python-dotenv>=1.0.0 